# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OurDay API Docs",
    "version": "1.0.0",
    "description": "OurDay API 문서"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://api.ourday.kr",
      "description": "Local Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "WeddingDetailWithId": {
        "type": "object",
        "properties": {
          "weddingId": {
            "type": "integer",
            "example": 1
          },
          "sections": {
            "$ref": "#/components/schemas/WeddingSections"
          },
          "sectionSettings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeddingSectionSetting"
            },
            "example": [
              {
                "sectionKey": "main",
                "isVisible": true,
                "displayOrder": 1
              },
              {
                "sectionKey": "shareLink",
                "isVisible": true,
                "displayOrder": 2
              },
              {
                "sectionKey": "weddingInfo",
                "isVisible": true,
                "displayOrder": 3
              },
              {
                "sectionKey": "familyInfo",
                "isVisible": true,
                "displayOrder": 4
              },
              {
                "sectionKey": "invitationMessage",
                "isVisible": true,
                "displayOrder": 5
              },
              {
                "sectionKey": "coupleIntro",
                "isVisible": true,
                "displayOrder": 6
              },
              {
                "sectionKey": "parentsIntro",
                "isVisible": true,
                "displayOrder": 7
              },
              {
                "sectionKey": "accountInfo",
                "isVisible": true,
                "displayOrder": 8
              },
              {
                "sectionKey": "locationInfo",
                "isVisible": true,
                "displayOrder": 9
              },
              {
                "sectionKey": "themeFont",
                "isVisible": true,
                "displayOrder": 10
              },
              {
                "sectionKey": "loadingScreen",
                "isVisible": true,
                "displayOrder": 11
              },
              {
                "sectionKey": "gallery",
                "isVisible": true,
                "displayOrder": 12
              },
              {
                "sectionKey": "flipbook",
                "isVisible": true,
                "displayOrder": 13
              }
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/weddings": {
      "get": {
        "summary": "나의 모든 청첩장 목록 조회",
        "description": "로그인한 사용자의 모든 청첩장(wedd) 리스트를 조회합니다.",
        "tags": [
          "Wedding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "messages": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "data": {
                      "type": "array",
                      "description": "wedd 테이블 리스트 (간단 정보)",
                      "items": {
                        "type": "object",
                        "properties": {
                          "weddingId": {
                            "type": "integer",
                            "example": 1
                          },
                          "weddingTitle": {
                            "type": "string",
                            "example": "관나의 청첩장"
                          },
                          "mainImageUrl": {
                            "type": "string",
                            "example": "/upload/weddingId/image.jpg"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "신규 청첩장 생성",
        "description": "새로운 청첩장을 생성하고 초기 섹션 및 섹션 설정을 함께 반환합니다.",
        "tags": [
          "Wedding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "messages": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "data": {
                      "$ref": "#/components/schemas/WeddingDetailWithId"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/weddings/{weddingId}": {
      "get": {
        "summary": "청첩장 상세 조회",
        "description": "특정 청첩장에 대한 모든 섹션 데이터 및 섹션 설정을 조회합니다.",
        "tags": [
          "Wedding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weddingId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "청첩장 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "messages": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "data": {
                      "$ref": "#/components/schemas/WeddingDetail"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "청첩장 전체 교체(섹션 전체 저장)",
        "description": "하나의 청첩장에 포함된 모든 섹션 데이터 및 섹션 설정을 한 번에 저장합니다.",
        "tags": [
          "Wedding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weddingId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeddingDetail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "교체 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "messages": {
                      "type": "string",
                      "nullable": true,
                      "example": "저장되었습니다."
                    },
                    "data": {
                      "$ref": "#/components/schemas/WeddingDetail"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "청첩장 삭제",
        "description": "해당 청첩장 및 관련 상세/섹션설정/미디어를 모두 삭제합니다.",
        "tags": [
          "Wedding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weddingId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "messages": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "data": {
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/weddings/{weddingId}/sections/settings": {
      "patch": {
        "summary": "섹션 표시 설정 및 순서 변경",
        "description": "섹션의 표시 여부(isVisible)와 순서를 일괄 수정합니다.",
        "tags": [
          "Wedding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "weddingId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeddingSectionSettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "섹션 설정 변경 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    },
                    "messages": {
                      "type": "string",
                      "nullable": true,
                      "example": "설정이 변경되었습니다."
                    },
                    "data": {
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}